+++ src/main/java/isg/pfe/travelAgency/Repositories/GuideRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface GuideRepository extends JpaRepository<Guide,Long> {
+public interface GuideRepository extends JpaRepository<Guide,Integer> {
 }
