+++ src/main/java/isg/pfe/travelAgency/Repositories/LocationRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import java.util.Optional;
 
 @RepositoryRestResource
-public interface LocationRepository extends JpaRepository<Location,Long> {
+public interface LocationRepository extends JpaRepository<Location,Integer> {
     Optional<Location> findByLatitudeAndLongitude(float Latitude, float Longitude);
 }
