+++ src/main/java/isg/pfe/travelAgency/Repositories/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import java.util.Optional;
 
 @RepositoryRestResource
-public interface UserRepository extends JpaRepository <UserU,Long> {
+public interface UserRepository extends JpaRepository <UserU,Integer> {
 
   Optional<UserU> findByUserName(String userName);
   UserU findByUserNameAndPassword(String userName,String Password);
