+++ src/main/java/isg/pfe/travelAgency/Repositories/VehicleModelRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface VehicleModelRepository extends JpaRepository<VehicleModel,Long> {
+public interface VehicleModelRepository extends JpaRepository<VehicleModel,Integer> {
 }
