+++ src/main/java/isg/pfe/travelAgency/Services/GuideServices.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface GuideServices {
     ResponseEntity<?> SaveGuide(Guide guide);
     ResponseEntity<List<Guide>> ListGuides();
-    ResponseEntity DeleteGuide(Long id);
-    ResponseEntity UpdateGuide(Long id, Guide newGuide);
-    ResponseEntity FindGuide(Long id);
+    ResponseEntity DeleteGuide(Integer id);
+    ResponseEntity UpdateGuide(Integer id, Guide newGuide);
+    ResponseEntity FindGuide(Integer id);
 }
