+++ src/main/java/isg/pfe/travelAgency/Services/LocationServices.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface LocationServices {
     ResponseEntity<?> SaveLocation(Location location);
     ResponseEntity<List<Location>> ListLocations();
-    ResponseEntity DeleteLocation(Long id);
-    ResponseEntity UpdateLocation(Long id,Location newLocation);
-    ResponseEntity FindLocation(Long id);
+    ResponseEntity DeleteLocation(Integer id);
+    ResponseEntity UpdateLocation(Integer id,Location newLocation);
+    ResponseEntity FindLocation(Integer id);
     ResponseEntity Optimize(List<Location> locations);
 }
