+++ src/main/java/isg/pfe/travelAgency/Services/TripServices.java
@@ -9,8 +9,8 @@ import java.util.List;
 public interface TripServices {
     ResponseEntity<?> SaveTrip(Trip trip);
     ResponseEntity<List<Trip>> ListTrip();
-    ResponseEntity DeleteTrip(Long id);
-    ResponseEntity UpdateTrip(Long id, Trip newTrip);
-    ResponseEntity FindTrip(Long id);
+    ResponseEntity DeleteTrip(Integer id);
+    ResponseEntity UpdateTrip(Integer id, Trip newTrip);
+    ResponseEntity FindTrip(Integer id);
     ResponseEntity <List<Trip>>FindTripByDriver(String id);
 }
